Move edit-paid-ads-campaign-form-content.js to shared components folder
